refactor: refactor test to ensure each test validates one thing.
